feat: add case unknown operation
