+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "type": "module",
   "scripts": {
-    "start": "node src/general/username.js",
+    "start": "node src/start/start.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
