+++ src/utils/redMessage.js
@@ -1,0 +1,3 @@
+export const redErrorMessage = (message, colorCode) => {
+    console.error(`\x1b[${colorCode}m${message}\x1b[0m`);
+};
